# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "description": "Authenticates a user with username and password and returns JWT access and refresh tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login with username and password",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revokes the current access token and optionally the refresh token by blacklisting their JTIs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout and revoke tokens",
                "parameters": [
                    {
                        "description": "Optional refresh token to also revoke",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "description": "Generates a new JWT access token using a valid refresh token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh access token",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New access token",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "description": "Creates a new user account and returns JWT access and refresh tokens. Registration must be enabled in server configuration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AccessKey": {
            "type": "object",
            "properties": {
                "access_key": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "description": "nil = never expires",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_used_at": {
                    "type": "string"
                },
                "name": {
                    "description": "Human label to tell keys apart",
                    "type": "string"
                },
                "read_only": {
                    "description": "Deny mutating S3 operations",
                    "type": "boolean"
                },
                "temporary": {
                    "description": "bkt-STS short-lived credential: excluded from the key limit, hard-deleted after expiry",
                    "type": "boolean"
                },
                "user": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.AuthResponse": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
        "models.Bucket": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "lifecycle": {
                    "description": "Lifecycle holds the bucket's lifecycle configuration as JSON (see\nLifecycleConfig); nil = no lifecycle rules.",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "objects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Object"
                    }
                },
                "owner": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "owner_id": {
                    "type": "string"
                },
                "quota_bytes": {
                    "description": "QuotaBytes caps the total size of CURRENT objects in the bucket\n(version storage is not counted). 0 = unlimited.",
                    "type": "integer"
                },
                "region": {
                    "type": "string"
                },
                "replicate_to": {
                    "description": "ReplicateTo mirrors this bucket's current objects into another bkt\nbucket (periodic sync; the target is fully managed by replication).",
                    "type": "string"
                },
                "retention_days": {
                    "description": "RetentionDays (WORM-lite): while set, no version of any object can be\npermanently removed until it is older than this many days. Requires\nversioning to be enabled. Plain deletes still create markers (data is\npreserved); version-addressed deletes, lifecycle purges, and bucket\ndeletion are refused for retained data. 0 = off.",
                    "type": "integer"
                },
                "s3_config": {
                    "$ref": "#/definitions/models.S3Configuration"
                },
                "s3_config_id": {
                    "description": "Optional: specific S3 config to use",
                    "type": "string"
                },
                "storage_backend": {
                    "description": "\"local\" or \"s3\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "versioning": {
                    "description": "Versioning state: \"\" (never enabled), \"enabled\", or \"suspended\".\nSuspended keeps existing versions browsable but stops creating new ones\n(a documented simplification of AWS's null-version semantics).",
                    "type": "string"
                },
                "webhook_events": {
                    "type": "string"
                },
                "webhook_url": {
                    "description": "Webhook event notifications: POSTed as JSON to WebhookURL for the event\ntypes in WebhookEvents (csv of \"created\",\"removed\"). Empty URL = off.",
                    "type": "string"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 1024
                },
                "username": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
        "models.Policy": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "document": {
                    "description": "IAM policy JSON",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "users": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "models.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "username": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                }
            }
        },
        "models.S3Configuration": {
            "type": "object",
            "properties": {
                "access_key_id": {
                    "type": "string"
                },
                "bucket_prefix": {
                    "type": "string"
                },
                "buckets": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Bucket"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "force_path_style": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "use_ssl": {
                    "type": "boolean"
                }
            }
        },
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "access_keys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AccessKey"
                    }
                },
                "buckets": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Bucket"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "is_locked": {
                    "description": "Account lock status",
                    "type": "boolean"
                },
                "policies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Policy"
                    }
                },
                "sso_email": {
                    "description": "Email from SSO (may differ from Email)",
                    "type": "string"
                },
                "sso_id": {
                    "description": "Unique ID from SSO provider",
                    "type": "string"
                },
                "sso_provider": {
                    "description": "SSO fields",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}